test(storage): extract peerstore inline tests to sibling file
Move the 20 Peerstore unit tests into peerstore_tests.rs without
changing production code or test behavior. Preserve module path
stores::peerstore::tests and private-item access via child module.
